requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase==2.0.3
python-dotenv==1.0.0
alpaca-py==0.43.2
pandas==2.3.3
numpy==2.4.0
streamlit>=1.29.0
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0
websockets==15.0.1
